+++ web/src/elements/commands/shared.ts
@@ -1,0 +1,6 @@
+export interface CommandPaletteCommand {
+    label: string;
+    description?: string;
+    group?: string;
+    action: () => void | Promise<void>;
+}
